+++ discord/ext/commands/flags.py
@@ -36,7 +36,7 @@ from discord.utils import resolve_annotation
 from .view import StringView
 from .converter import run_converters
 
-from discord.utils import maybe_coroutine
+from discord.utils import maybe_coroutine, MISSING
 from dataclasses import dataclass, field
 from typing import (
     Dict,
@@ -67,14 +67,6 @@ __all__ = (
 
 if TYPE_CHECKING:
     from .context import Context
-
-
-class _MissingSentinel:
-    def __repr__(self):
-        return 'MISSING'
-
-
-MISSING: Any = _MissingSentinel()
 
 
 @dataclass
